Fixed imgui user config include for other libraries; Fixed imgui blocking cursor setting for RmlUI; Fixed shipping build; Removed deprecated text renderer and sprite renderer; Moved Context container code directly into GUI extension;
