+++ Engine/source/EtGUI/ImGui/ImGuiUtil.h
@@ -2,9 +2,10 @@
 #include <imgui/imgui.h>
 
 
-#ifndef IMGUI_DISABLE
+#if ET_IMGUI_ENABLED
 
 #include <EtCore/Input/KeyCodes.h>
+#include <EtCore/Util/CursorShapes.h>
 
 
 namespace et {
@@ -15,6 +16,7 @@ namespace ImguiUtil {
 
 ImGuiMouseButton GetButton(E_MouseButton const button);
 ImGuiKey GetKey(E_KbdKey const key);
+core::E_CursorShape GetCursorShape(ImGuiMouseCursor const cursor);
 
 
 // Converting et::math types to Rml counterparts
@@ -30,4 +32,4 @@ vec2 ToEtm(ImVec2 const& rhs);
 } // namespace et
 
 
-#endif // ndef IMGUI_DISABLE
+#endif // ET_IMGUI_ENABLED
