+++ Engine/source/EtGUI/Rendering/SceneRendererGUI.h
@@ -1,6 +1,4 @@
 #pragma once
-#include "TextRenderer.h"
-#include "SpriteRenderer.h"
 #include "GuiRenderer.h"
 
 #include <EtGUI/GuiExtension.h>
@@ -33,30 +31,22 @@ public:
 
 	// functionality
 	//---------------
+private:
 	void DrawInWorld(render::T_FbLoc const targetFb, GuiExtension& guiExt, core::slot_map<mat4> const& nodes);
-	void DrawOverlay(render::T_FbLoc const targetFb, GuiExtension& guiExt, core::slot_map<mat4> const& nodes);
+	void DrawOverlay(render::T_FbLoc const targetFb, GuiExtension& guiExt);
 
-	// accessors
-	//-----------
-	TextRenderer& GetTextRenderer() { return m_TextRenderer; }
-	SpriteRenderer& GetSpriteRenderer() { return m_SpriteRenderer; }
 
 	// Data
 	///////
 
-private:
 	bool m_IsInitialized = false;
+
+	GuiRenderer m_GuiRenderer;
 
 	// render event hooks
 	Ptr<render::T_RenderEventDispatcher> m_EventDispatcher;
 	render::T_RenderEventCallbackId m_WorldCallbackId = render::T_RenderEventDispatcher::INVALID_ID;
 	render::T_RenderEventCallbackId m_OverlayCallbackId = render::T_RenderEventDispatcher::INVALID_ID;
-
-	// renderers
-	//-----------
-	TextRenderer m_TextRenderer;
-	SpriteRenderer m_SpriteRenderer;
-	GuiRenderer m_GuiRenderer;
 };
