+++ Projects/Demo/source/Runtime/MainFramework.h
@@ -37,9 +37,10 @@ private:
 
 	size_t m_CurrentScene = 0u;
 
+#	ifndef IMGUI_DISABLE
 	bool m_DrawDebugInfo = true;
 	bool m_DrawDemoWindow = false;
-	bool m_ShowGui = true;
+#	endif
 };
